SocketIOClient.connect() should fire timeout handler if not connected
Previously, the timeout handler was only being invoked if the status was
not `.connected` or `.disconnected`. It's important that it doesn't
change its status again if it's already disconnected (there's some code
inside the `didSet` closure of the property), but it's also important
that the `handler` gets invoked. If this doesn't happen and multiple
calls are made to `connect()` without a successful connection, only the
first of these will fire its timeout handler... the rest will continue
to hang indefinitely.